Move FdTable to public location, fix up imports
